move TeamChat to it's own component
